+++ internal/operation/storage.go
@@ -7,5 +7,5 @@ type Storage interface {
 	FindOne(ctx context.Context, id string) (Operation, error)
 	Update(ctx context.Context, op Operation) error
 	GetOperationsByExpression(ctx context.Context, exprID string) ([]Operation, error)
-	UpdateOperationResult(ctx context.Context, id string, result float32, status string) error
+	UpdateOperationResult(ctx context.Context, id string, result float64, status string) error
 }
